Refactor setup to setupJasmine to support both jasmine-jest and jasmine v3 environments; Add jasmine tests
