--- __recordings__/polly-jest-integration_3899455233/should-configure-polly-when-needed_241621399/recording.har
@@ -1,137 +1,0 @@
-{
-  "log": {
-    "_recordingName": "polly jest integration/should configure polly when needed",
-    "creator": {
-      "comment": "persister:fs",
-      "name": "Polly.JS",
-      "version": "1.2.0"
-    },
-    "entries": [
-      {
-        "_id": "dfe8c63deff866306260469b7311c301",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 0,
-          "cookies": [],
-          "headers": [],
-          "headersSize": 69,
-          "httpVersion": "HTTP/1.1",
-          "method": "GET",
-          "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/-999999"
-        },
-        "response": {
-          "bodySize": 2,
-          "content": {
-            "mimeType": "application/json; charset=utf-8",
-            "size": 2,
-            "text": "{}"
-          },
-          "cookies": [
-            {
-              "domain": ".typicode.com",
-              "expires": "2019-10-06T13:01:26.000Z",
-              "httpOnly": true,
-              "name": "__cfduid",
-              "path": "/",
-              "value": "d377cd4804d64ff867de54209a9dde8ae1538830886"
-            }
-          ],
-          "headers": [
-            {
-              "name": "access-control-allow-credentials",
-              "value": "true"
-            },
-            {
-              "name": "cache-control",
-              "value": "public, max-age=14400"
-            },
-            {
-              "name": "cf-cache-status",
-              "value": "EXPIRED"
-            },
-            {
-              "name": "cf-ray",
-              "value": "4658510dba06233c-FRA"
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "content-length",
-              "value": "2"
-            },
-            {
-              "name": "content-type",
-              "value": "application/json; charset=utf-8"
-            },
-            {
-              "name": "date",
-              "value": "Sat, 06 Oct 2018 13:01:26 GMT"
-            },
-            {
-              "name": "etag",
-              "value": "W/\"2-vyGp6PvFo4RvsFtPoIWeCReyIC8\""
-            },
-            {
-              "name": "expect-ct",
-              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
-            },
-            {
-              "name": "expires",
-              "value": "Sat, 06 Oct 2018 17:01:26 GMT"
-            },
-            {
-              "name": "pragma",
-              "value": "no-cache"
-            },
-            {
-              "name": "server",
-              "value": "cloudflare"
-            },
-            {
-              "name": "set-cookie",
-              "value": "__cfduid=d377cd4804d64ff867de54209a9dde8ae1538830886; expires=Sun, 06-Oct-19 13:01:26 GMT; path=/; domain=.typicode.com; HttpOnly"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, Accept-Encoding"
-            },
-            {
-              "name": "via",
-              "value": "1.1 vegur"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "x-powered-by",
-              "value": "Express"
-            }
-          ],
-          "headersSize": 723,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 404,
-          "statusText": "Not Found"
-        },
-        "startedDateTime": "2018-10-06T13:01:25.953Z",
-        "time": 407,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 407
-        }
-      }
-    ],
-    "pages": [],
-    "version": "1.2"
-  }
-}
