--- __recordings__/polly-jest-integration_3899455233/should-have-polly-running-with-test-_2577911815/recording.har
@@ -1,141 +1,0 @@
-{
-  "log": {
-    "_recordingName": "polly jest integration/should have polly running with `test`",
-    "creator": {
-      "comment": "persister:fs",
-      "name": "Polly.JS",
-      "version": "1.2.0"
-    },
-    "entries": [
-      {
-        "_id": "ffbc4836d419fc265c3b85cbe1b7f22e",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 0,
-          "cookies": [],
-          "headers": [],
-          "headersSize": 63,
-          "httpVersion": "HTTP/1.1",
-          "method": "GET",
-          "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/1"
-        },
-        "response": {
-          "bodySize": 292,
-          "content": {
-            "mimeType": "application/json; charset=utf-8",
-            "size": 292,
-            "text": "{\n  \"userId\": 1,\n  \"id\": 1,\n  \"title\": \"sunt aut facere repellat provident occaecati excepturi optio reprehenderit\",\n  \"body\": \"quia et suscipit\\nsuscipit recusandae consequuntur expedita et cum\\nreprehenderit molestiae ut ut quas totam\\nnostrum rerum est autem sunt rem eveniet architecto\"\n}"
-          },
-          "cookies": [
-            {
-              "domain": ".typicode.com",
-              "expires": "2019-10-06T13:01:25.000Z",
-              "httpOnly": true,
-              "name": "__cfduid",
-              "path": "/",
-              "value": "dae37d9bf205af784db2a40e63c6d98e31538830885"
-            }
-          ],
-          "headers": [
-            {
-              "name": "access-control-allow-credentials",
-              "value": "true"
-            },
-            {
-              "name": "cache-control",
-              "value": "public, max-age=14400"
-            },
-            {
-              "name": "cf-cache-status",
-              "value": "HIT"
-            },
-            {
-              "name": "cf-ray",
-              "value": "46585109faae9756-FRA"
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "content-encoding",
-              "value": "gzip"
-            },
-            {
-              "name": "content-type",
-              "value": "application/json; charset=utf-8"
-            },
-            {
-              "name": "date",
-              "value": "Sat, 06 Oct 2018 13:01:25 GMT"
-            },
-            {
-              "name": "etag",
-              "value": "W/\"124-yiKdLzqO5gfBrJFrcdJ8Yq0LGnU\""
-            },
-            {
-              "name": "expect-ct",
-              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
-            },
-            {
-              "name": "expires",
-              "value": "Sat, 06 Oct 2018 17:01:25 GMT"
-            },
-            {
-              "name": "pragma",
-              "value": "no-cache"
-            },
-            {
-              "name": "server",
-              "value": "cloudflare"
-            },
-            {
-              "name": "set-cookie",
-              "value": "__cfduid=dae37d9bf205af784db2a40e63c6d98e31538830885; expires=Sun, 06-Oct-19 13:01:25 GMT; path=/; domain=.typicode.com; HttpOnly"
-            },
-            {
-              "name": "transfer-encoding",
-              "value": "chunked"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, Accept-Encoding"
-            },
-            {
-              "name": "via",
-              "value": "1.1 vegur"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "x-powered-by",
-              "value": "Express"
-            }
-          ],
-          "headersSize": 754,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 200,
-          "statusText": "OK"
-        },
-        "startedDateTime": "2018-10-06T13:01:25.346Z",
-        "time": 102,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 102
-        }
-      }
-    ],
-    "pages": [],
-    "version": "1.2"
-  }
-}
