--- __recordings__/polly-jest-integration_3899455233/should-work-with-each-capturing-recording-for-post-2_1192146480/recording.har
@@ -1,141 +1,0 @@
-{
-  "log": {
-    "_recordingName": "polly jest integration/should work with each, capturing recording for post #2",
-    "creator": {
-      "comment": "persister:fs",
-      "name": "Polly.JS",
-      "version": "1.2.0"
-    },
-    "entries": [
-      {
-        "_id": "be0c17ae6afbec746230f8e054c63e36",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 0,
-          "cookies": [],
-          "headers": [],
-          "headersSize": 63,
-          "httpVersion": "HTTP/1.1",
-          "method": "GET",
-          "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/2"
-        },
-        "response": {
-          "bodySize": 278,
-          "content": {
-            "mimeType": "application/json; charset=utf-8",
-            "size": 278,
-            "text": "{\n  \"userId\": 1,\n  \"id\": 2,\n  \"title\": \"qui est esse\",\n  \"body\": \"est rerum tempore vitae\\nsequi sint nihil reprehenderit dolor beatae ea dolores neque\\nfugiat blanditiis voluptate porro vel nihil molestiae ut reiciendis\\nqui aperiam non debitis possimus qui neque nisi nulla\"\n}"
-          },
-          "cookies": [
-            {
-              "domain": ".typicode.com",
-              "expires": "2019-10-06T13:01:25.000Z",
-              "httpOnly": true,
-              "name": "__cfduid",
-              "path": "/",
-              "value": "d19c75624b188ec11266e7a9b9410c4a11538830885"
-            }
-          ],
-          "headers": [
-            {
-              "name": "access-control-allow-credentials",
-              "value": "true"
-            },
-            {
-              "name": "cache-control",
-              "value": "public, max-age=14400"
-            },
-            {
-              "name": "cf-cache-status",
-              "value": "HIT"
-            },
-            {
-              "name": "cf-ray",
-              "value": "4658510bdb8f6373-FRA"
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "content-encoding",
-              "value": "gzip"
-            },
-            {
-              "name": "content-type",
-              "value": "application/json; charset=utf-8"
-            },
-            {
-              "name": "date",
-              "value": "Sat, 06 Oct 2018 13:01:25 GMT"
-            },
-            {
-              "name": "etag",
-              "value": "W/\"116-jnDuMpjju89+9j7e0BqkdFsVRjs\""
-            },
-            {
-              "name": "expect-ct",
-              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
-            },
-            {
-              "name": "expires",
-              "value": "Sat, 06 Oct 2018 17:01:25 GMT"
-            },
-            {
-              "name": "pragma",
-              "value": "no-cache"
-            },
-            {
-              "name": "server",
-              "value": "cloudflare"
-            },
-            {
-              "name": "set-cookie",
-              "value": "__cfduid=d19c75624b188ec11266e7a9b9410c4a11538830885; expires=Sun, 06-Oct-19 13:01:25 GMT; path=/; domain=.typicode.com; HttpOnly"
-            },
-            {
-              "name": "transfer-encoding",
-              "value": "chunked"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, Accept-Encoding"
-            },
-            {
-              "name": "via",
-              "value": "1.1 vegur"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "x-powered-by",
-              "value": "Express"
-            }
-          ],
-          "headersSize": 754,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 200,
-          "statusText": "OK"
-        },
-        "startedDateTime": "2018-10-06T13:01:25.647Z",
-        "time": 92,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 92
-        }
-      }
-    ],
-    "pages": [],
-    "version": "1.2"
-  }
-}
