--- __recordings__/polly-jest-integration_3899455233/should-work-with-each-capturing-recording-for-post-3_1208924099/recording.har
@@ -1,141 +1,0 @@
-{
-  "log": {
-    "_recordingName": "polly jest integration/should work with each, capturing recording for post #3",
-    "creator": {
-      "comment": "persister:fs",
-      "name": "Polly.JS",
-      "version": "1.2.0"
-    },
-    "entries": [
-      {
-        "_id": "5e71aaa62033415e38b0745f1e0bb797",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 0,
-          "cookies": [],
-          "headers": [],
-          "headersSize": 63,
-          "httpVersion": "HTTP/1.1",
-          "method": "GET",
-          "queryString": [],
-          "url": "https://jsonplaceholder.typicode.com/posts/3"
-        },
-        "response": {
-          "bodySize": 283,
-          "content": {
-            "mimeType": "application/json; charset=utf-8",
-            "size": 283,
-            "text": "{\n  \"userId\": 1,\n  \"id\": 3,\n  \"title\": \"ea molestias quasi exercitationem repellat qui ipsa sit aut\",\n  \"body\": \"et iusto sed quo iure\\nvoluptatem occaecati omnis eligendi aut ad\\nvoluptatem doloribus vel accusantium quis pariatur\\nmolestiae porro eius odio et labore et velit aut\"\n}"
-          },
-          "cookies": [
-            {
-              "domain": ".typicode.com",
-              "expires": "2019-10-06T13:01:25.000Z",
-              "httpOnly": true,
-              "name": "__cfduid",
-              "path": "/",
-              "value": "dc99698e75a31945dd0fd58e0dc27fae71538830885"
-            }
-          ],
-          "headers": [
-            {
-              "name": "access-control-allow-credentials",
-              "value": "true"
-            },
-            {
-              "name": "cache-control",
-              "value": "public, max-age=14400"
-            },
-            {
-              "name": "cf-cache-status",
-              "value": "HIT"
-            },
-            {
-              "name": "cf-ray",
-              "value": "4658510c68c2978c-FRA"
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "content-encoding",
-              "value": "gzip"
-            },
-            {
-              "name": "content-type",
-              "value": "application/json; charset=utf-8"
-            },
-            {
-              "name": "date",
-              "value": "Sat, 06 Oct 2018 13:01:25 GMT"
-            },
-            {
-              "name": "etag",
-              "value": "W/\"11b-USacuIw5a/iXAGdNKBvqr/TbMTc\""
-            },
-            {
-              "name": "expect-ct",
-              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
-            },
-            {
-              "name": "expires",
-              "value": "Sat, 06 Oct 2018 17:01:25 GMT"
-            },
-            {
-              "name": "pragma",
-              "value": "no-cache"
-            },
-            {
-              "name": "server",
-              "value": "cloudflare"
-            },
-            {
-              "name": "set-cookie",
-              "value": "__cfduid=dc99698e75a31945dd0fd58e0dc27fae71538830885; expires=Sun, 06-Oct-19 13:01:25 GMT; path=/; domain=.typicode.com; HttpOnly"
-            },
-            {
-              "name": "transfer-encoding",
-              "value": "chunked"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, Accept-Encoding"
-            },
-            {
-              "name": "via",
-              "value": "1.1 vegur"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "x-powered-by",
-              "value": "Express"
-            }
-          ],
-          "headersSize": 754,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 200,
-          "statusText": "OK"
-        },
-        "startedDateTime": "2018-10-06T13:01:25.743Z",
-        "time": 107,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 107
-        }
-      }
-    ],
-    "pages": [],
-    "version": "1.2"
-  }
-}
