+++ jest/__recordings__/setupPolly_466486458/should-configure-polly-when-needed_241621399/recording.har
@@ -1,0 +1,137 @@
+{
+  "log": {
+    "_recordingName": "setupPolly/should configure polly when needed",
+    "creator": {
+      "comment": "persister:fs",
+      "name": "Polly.JS",
+      "version": "1.2.0"
+    },
+    "entries": [
+      {
+        "_id": "dfe8c63deff866306260469b7311c301",
+        "_order": 0,
+        "cache": {},
+        "request": {
+          "bodySize": 0,
+          "cookies": [],
+          "headers": [],
+          "headersSize": 69,
+          "httpVersion": "HTTP/1.1",
+          "method": "GET",
+          "queryString": [],
+          "url": "https://jsonplaceholder.typicode.com/posts/-999999"
+        },
+        "response": {
+          "bodySize": 2,
+          "content": {
+            "mimeType": "application/json; charset=utf-8",
+            "size": 2,
+            "text": "{}"
+          },
+          "cookies": [
+            {
+              "domain": ".typicode.com",
+              "expires": "2019-10-13T09:17:30.000Z",
+              "httpOnly": true,
+              "name": "__cfduid",
+              "path": "/",
+              "value": "d11592fc4672926f51920ae767d14a7e81539422250"
+            }
+          ],
+          "headers": [
+            {
+              "name": "access-control-allow-credentials",
+              "value": "true"
+            },
+            {
+              "name": "cache-control",
+              "value": "public, max-age=14400"
+            },
+            {
+              "name": "cf-cache-status",
+              "value": "HIT"
+            },
+            {
+              "name": "cf-ray",
+              "value": "4690b6a9caa4639d-FRA"
+            },
+            {
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "name": "content-length",
+              "value": "2"
+            },
+            {
+              "name": "content-type",
+              "value": "application/json; charset=utf-8"
+            },
+            {
+              "name": "date",
+              "value": "Sat, 13 Oct 2018 09:17:30 GMT"
+            },
+            {
+              "name": "etag",
+              "value": "W/\"2-vyGp6PvFo4RvsFtPoIWeCReyIC8\""
+            },
+            {
+              "name": "expect-ct",
+              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
+            },
+            {
+              "name": "expires",
+              "value": "Sat, 13 Oct 2018 13:17:30 GMT"
+            },
+            {
+              "name": "pragma",
+              "value": "no-cache"
+            },
+            {
+              "name": "server",
+              "value": "cloudflare"
+            },
+            {
+              "name": "set-cookie",
+              "value": "__cfduid=d11592fc4672926f51920ae767d14a7e81539422250; expires=Sun, 13-Oct-19 09:17:30 GMT; path=/; domain=.typicode.com; HttpOnly"
+            },
+            {
+              "name": "vary",
+              "value": "Origin, Accept-Encoding"
+            },
+            {
+              "name": "via",
+              "value": "1.1 vegur"
+            },
+            {
+              "name": "x-content-type-options",
+              "value": "nosniff"
+            },
+            {
+              "name": "x-powered-by",
+              "value": "Express"
+            }
+          ],
+          "headersSize": 719,
+          "httpVersion": "HTTP/1.1",
+          "redirectURL": "",
+          "status": 404,
+          "statusText": "Not Found"
+        },
+        "startedDateTime": "2018-10-13T09:17:30.432Z",
+        "time": 69,
+        "timings": {
+          "blocked": -1,
+          "connect": -1,
+          "dns": -1,
+          "receive": 0,
+          "send": 0,
+          "ssl": -1,
+          "wait": 69
+        }
+      }
+    ],
+    "pages": [],
+    "version": "1.2"
+  }
+}
