+++ jest/__recordings__/setupPolly_466486458/should-have-polly-running-with-it-_2136411752/recording.har
@@ -1,0 +1,141 @@
+{
+  "log": {
+    "_recordingName": "setupPolly/should have polly running with `it`",
+    "creator": {
+      "comment": "persister:fs",
+      "name": "Polly.JS",
+      "version": "1.2.0"
+    },
+    "entries": [
+      {
+        "_id": "be0c17ae6afbec746230f8e054c63e36",
+        "_order": 0,
+        "cache": {},
+        "request": {
+          "bodySize": 0,
+          "cookies": [],
+          "headers": [],
+          "headersSize": 63,
+          "httpVersion": "HTTP/1.1",
+          "method": "GET",
+          "queryString": [],
+          "url": "https://jsonplaceholder.typicode.com/posts/2"
+        },
+        "response": {
+          "bodySize": 278,
+          "content": {
+            "mimeType": "application/json; charset=utf-8",
+            "size": 278,
+            "text": "{\n  \"userId\": 1,\n  \"id\": 2,\n  \"title\": \"qui est esse\",\n  \"body\": \"est rerum tempore vitae\\nsequi sint nihil reprehenderit dolor beatae ea dolores neque\\nfugiat blanditiis voluptate porro vel nihil molestiae ut reiciendis\\nqui aperiam non debitis possimus qui neque nisi nulla\"\n}"
+          },
+          "cookies": [
+            {
+              "domain": ".typicode.com",
+              "expires": "2019-10-13T09:17:30.000Z",
+              "httpOnly": true,
+              "name": "__cfduid",
+              "path": "/",
+              "value": "d8bfadf73b0f304c22fba02d0797f1c5d1539422250"
+            }
+          ],
+          "headers": [
+            {
+              "name": "access-control-allow-credentials",
+              "value": "true"
+            },
+            {
+              "name": "cache-control",
+              "value": "public, max-age=14400"
+            },
+            {
+              "name": "cf-cache-status",
+              "value": "HIT"
+            },
+            {
+              "name": "cf-ray",
+              "value": "4690b6a798022744-FRA"
+            },
+            {
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "name": "content-encoding",
+              "value": "gzip"
+            },
+            {
+              "name": "content-type",
+              "value": "application/json; charset=utf-8"
+            },
+            {
+              "name": "date",
+              "value": "Sat, 13 Oct 2018 09:17:30 GMT"
+            },
+            {
+              "name": "etag",
+              "value": "W/\"116-jnDuMpjju89+9j7e0BqkdFsVRjs\""
+            },
+            {
+              "name": "expect-ct",
+              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
+            },
+            {
+              "name": "expires",
+              "value": "Sat, 13 Oct 2018 13:17:30 GMT"
+            },
+            {
+              "name": "pragma",
+              "value": "no-cache"
+            },
+            {
+              "name": "server",
+              "value": "cloudflare"
+            },
+            {
+              "name": "set-cookie",
+              "value": "__cfduid=d8bfadf73b0f304c22fba02d0797f1c5d1539422250; expires=Sun, 13-Oct-19 09:17:30 GMT; path=/; domain=.typicode.com; HttpOnly"
+            },
+            {
+              "name": "transfer-encoding",
+              "value": "chunked"
+            },
+            {
+              "name": "vary",
+              "value": "Origin, Accept-Encoding"
+            },
+            {
+              "name": "via",
+              "value": "1.1 vegur"
+            },
+            {
+              "name": "x-content-type-options",
+              "value": "nosniff"
+            },
+            {
+              "name": "x-powered-by",
+              "value": "Express"
+            }
+          ],
+          "headersSize": 754,
+          "httpVersion": "HTTP/1.1",
+          "redirectURL": "",
+          "status": 200,
+          "statusText": "OK"
+        },
+        "startedDateTime": "2018-10-13T09:17:30.086Z",
+        "time": 76,
+        "timings": {
+          "blocked": -1,
+          "connect": -1,
+          "dns": -1,
+          "receive": 0,
+          "send": 0,
+          "ssl": -1,
+          "wait": 76
+        }
+      }
+    ],
+    "pages": [],
+    "version": "1.2"
+  }
+}
