+++ jest/__recordings__/setupPolly_466486458/should-have-polly-running-with-test-_2577911815/recording.har
@@ -1,0 +1,141 @@
+{
+  "log": {
+    "_recordingName": "setupPolly/should have polly running with `test`",
+    "creator": {
+      "comment": "persister:fs",
+      "name": "Polly.JS",
+      "version": "1.2.0"
+    },
+    "entries": [
+      {
+        "_id": "ffbc4836d419fc265c3b85cbe1b7f22e",
+        "_order": 0,
+        "cache": {},
+        "request": {
+          "bodySize": 0,
+          "cookies": [],
+          "headers": [],
+          "headersSize": 63,
+          "httpVersion": "HTTP/1.1",
+          "method": "GET",
+          "queryString": [],
+          "url": "https://jsonplaceholder.typicode.com/posts/1"
+        },
+        "response": {
+          "bodySize": 292,
+          "content": {
+            "mimeType": "application/json; charset=utf-8",
+            "size": 292,
+            "text": "{\n  \"userId\": 1,\n  \"id\": 1,\n  \"title\": \"sunt aut facere repellat provident occaecati excepturi optio reprehenderit\",\n  \"body\": \"quia et suscipit\\nsuscipit recusandae consequuntur expedita et cum\\nreprehenderit molestiae ut ut quas totam\\nnostrum rerum est autem sunt rem eveniet architecto\"\n}"
+          },
+          "cookies": [
+            {
+              "domain": ".typicode.com",
+              "expires": "2019-10-13T09:17:30.000Z",
+              "httpOnly": true,
+              "name": "__cfduid",
+              "path": "/",
+              "value": "da54b0a02b09e5911cdb902d44b4a1c951539422250"
+            }
+          ],
+          "headers": [
+            {
+              "name": "access-control-allow-credentials",
+              "value": "true"
+            },
+            {
+              "name": "cache-control",
+              "value": "public, max-age=14400"
+            },
+            {
+              "name": "cf-cache-status",
+              "value": "HIT"
+            },
+            {
+              "name": "cf-ray",
+              "value": "4690b6a71cf16355-FRA"
+            },
+            {
+              "name": "connection",
+              "value": "close"
+            },
+            {
+              "name": "content-encoding",
+              "value": "gzip"
+            },
+            {
+              "name": "content-type",
+              "value": "application/json; charset=utf-8"
+            },
+            {
+              "name": "date",
+              "value": "Sat, 13 Oct 2018 09:17:30 GMT"
+            },
+            {
+              "name": "etag",
+              "value": "W/\"124-yiKdLzqO5gfBrJFrcdJ8Yq0LGnU\""
+            },
+            {
+              "name": "expect-ct",
+              "value": "max-age=604800, report-uri=\"https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct\""
+            },
+            {
+              "name": "expires",
+              "value": "Sat, 13 Oct 2018 13:17:30 GMT"
+            },
+            {
+              "name": "pragma",
+              "value": "no-cache"
+            },
+            {
+              "name": "server",
+              "value": "cloudflare"
+            },
+            {
+              "name": "set-cookie",
+              "value": "__cfduid=da54b0a02b09e5911cdb902d44b4a1c951539422250; expires=Sun, 13-Oct-19 09:17:30 GMT; path=/; domain=.typicode.com; HttpOnly"
+            },
+            {
+              "name": "transfer-encoding",
+              "value": "chunked"
+            },
+            {
+              "name": "vary",
+              "value": "Origin, Accept-Encoding"
+            },
+            {
+              "name": "via",
+              "value": "1.1 vegur"
+            },
+            {
+              "name": "x-content-type-options",
+              "value": "nosniff"
+            },
+            {
+              "name": "x-powered-by",
+              "value": "Express"
+            }
+          ],
+          "headersSize": 754,
+          "httpVersion": "HTTP/1.1",
+          "redirectURL": "",
+          "status": 200,
+          "statusText": "OK"
+        },
+        "startedDateTime": "2018-10-13T09:17:29.942Z",
+        "time": 135,
+        "timings": {
+          "blocked": -1,
+          "connect": -1,
+          "dns": -1,
+          "receive": 0,
+          "send": 0,
+          "ssl": -1,
+          "wait": 135
+        }
+      }
+    ],
+    "pages": [],
+    "version": "1.2"
+  }
+}
