+++ package.json
@@ -5,8 +5,10 @@
   "main": "index.js",
   "scripts": {
     "build": "babel src --out-dir lib",
-    "test": "npm run build && jest",
-    "prepublishOnly": "npm run build"
+    "test:jest": "jest",
+    "test:jasmine": "NODE_ENV=test jasmine",
+    "prepare": "npm run build && npm run test:jest && npm run test:jasmine",
+    "prepublishOnly": "npm run prepare"
   },
   "files": [
     "LICENSE",
@@ -36,11 +38,17 @@
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^23.6.0",
     "eslint": "^5.6.1",
+    "jasmine": "^3.2.0",
     "jest": "^23.6.0",
     "node-fetch": "^2.2.0",
     "prettier": "^1.14.3"
   },
   "peerDependencies": {
     "@pollyjs/core": "*"
+  },
+  "jest": {
+    "testMatch": [
+      "<rootDir>/jest/**/*.test.js"
+    ]
   }
 }
